refactoring services for views and forms
